+++ tests/test_parser/test_tense_and_person.py
@@ -1,6 +1,7 @@
 from lxml import etree
 
-from verbecc.src.parsers.tense_template import TenseTemplate
+from verbecc.src.parsers.tense_template_parser import TenseTemplateParser
+from verbecc.src.defs.types.data.tense_template import TenseTemplate
 from verbecc.src.defs.types.lang_code import LangCodeISO639_1 as Lang
 from verbecc.src.defs.types.tense import Tense
 
@@ -17,9 +18,9 @@ def test_tense_and_person():
         </présent>"""
     tense_elem: etree._Element = etree.fromstring(tense_elem_str)
     tense = Tense.fr.Présent
-    tense_template = TenseTemplate(Lang.fr, mood, tense_elem)
+    tense_template = TenseTemplateParser(Lang.fr, mood).parse(tense_elem)
     assert tense_template.mood == mood
-    assert tense_template.name == str(tense.value)
+    assert tense_template.tense == tense
     assert tense_template.person_endings[0].get_ending() == "ie"
     assert tense_template.person_endings[0].get_alternate_ending_if_available() == "ye"
     assert tense_template.person_endings[0].get_person() == "1s"