+++ verbecc/src/defs/types/data/person_ending.py
@@ -1,37 +1,25 @@
-from lxml import etree
 from typing import List
 
+from verbecc.src.defs.types.data.element import Element
 from verbecc.src.defs.types.person import Person
 
 
-class PersonEnding:
+class PersonEnding(Element):
     """
-    p_elem
-    aka <p>
     Ending for a specific verb template, mood, tense and grammatical person
-    May also have an alternate ending for an alternative spelling.
-    E.g. Ending for aim:er indicatif présent 2nd Person Plural = ['ez']
-    E.g. Ending for pa:yer indicatif présent 1st Person Singular = ['ie', 'ye']
+    May also have one or more alternate endings for an alternative spellings
+    or regional variations.
+    E.g. Endings for aim:er indicatif présent 2nd Person Plural = ['ez']
+    E.g. Endings for pa:yer indicatif présent 1st Person Singular = ['ie', 'ye']
     Explanation: 'ye' is an alternate spelling (je paie, je paye)
-    p_elem
-        Example p_elems:
-            <p><i>ez</i></p>
-            <p><i>eoir</i><i>oir</i></p>
-            <p></p>
-
     person
     A grammar_defines.PERSONS value indicating which person
     this PersonEnding is for, e.g. for aim:er, "ez" is '2p' (second person plural)
     """
 
-    def __init__(self, p_elem: etree._Element, person: Person) -> None:
+    def __init__(self, person: Person, endings: List[str]) -> None:
         self.person = person
-        self.endings: List[str] = []
-        for i_elem in p_elem.findall("i", None):
-            ending = str("")
-            if i_elem.text is not None:
-                ending += str(i_elem.text)
-            self.endings.append(ending)
+        self.endings = endings
 
     def get_person(self) -> Person:
         return self.person