+++ verbecc/src/inflectors/inflector.py
@@ -1,9 +1,6 @@
 import logging
 
 from verbecc.src.defs.constants.config import DEVEL_MODE
-from verbecc.src.defs.types.mood import MoodEn as Mood
-from verbecc.src.defs.types.tense import TenseEn as Tense
-from verbecc.src.parsers.person_ending import PersonEnding
 
 logging_level = logging.CRITICAL + 1  # effectively disables logging
 if DEVEL_MODE:
@@ -20,20 +17,22 @@ logger = logging.getLogger(__name__)
 from abc import ABC, abstractmethod
 from typing import Dict, List, Tuple
 
-from verbecc.src.defs.types.gender import Gender
-from verbecc.src.defs.constants.grammar_defines import PARTICIPLE_INFLECTIONS
-from verbecc.src.defs.types.exceptions import ConjugatorError
-from verbecc.src.defs.types.lang_code import LangCodeISO639_1
-from verbecc.src.defs.types.partiple_inflection import ParticipleInflection
-from verbecc.src.defs.types.person import Person, is_singular
-from verbecc.src.defs.types.lang_specific_options import (
-    LangSpecificOptions,
-)
-from verbecc.src.parsers.conjugations_parser import ConjugationsParser
-from verbecc.src.parsers.conjugation_template import ConjugationTemplate
-from verbecc.src.parsers.tense_template import TenseTemplate
 from verbecc.src.conjugator.conjugation_object import ConjugationObjects
-from verbecc.src.parsers.verb import Verb
+from verbecc.src.defs.constants.grammar_defines import PARTICIPLE_INFLECTIONS
+from verbecc.src.defs.types.data.conjugation_template import ConjugationTemplate
+from verbecc.src.defs.types.data.person_ending import PersonEnding
+from verbecc.src.defs.types.data.tense_template import TenseTemplate
+from verbecc.src.defs.types.data.verb import Verb
+from verbecc.src.defs.types.data.verbs import Verbs
+from verbecc.src.defs.types.exceptions import ConjugatorError
+from verbecc.src.defs.types.gender import Gender
+from verbecc.src.defs.types.lang_code import LangCodeISO639_1
+from verbecc.src.defs.types.lang_specific_options import LangSpecificOptions
+from verbecc.src.defs.types.mood import MoodEn as Mood
+from verbecc.src.defs.types.participle_inflection import ParticipleInflection
+from verbecc.src.defs.types.person import Person, is_singular
+from verbecc.src.defs.types.tense import TenseEn as Tense
+from verbecc.src.parsers.conjugations_parser import ConjugationsParser
 from verbecc.src.parsers.verbs_parser import VerbsParser
 
 
@@ -46,8 +45,8 @@ class Inflector(ABC):
     # public:
 
     def __init__(self) -> None:
-        self._verb_parser = VerbsParser(self.lang)
-        self._conj_parser = ConjugationsParser(self.lang)
+        self._verbs: Verbs = VerbsParser(self.lang).parse()
+        self._conjugations = ConjugationsParser(self.lang).parse()
 
     @property
     @abstractmethod
@@ -55,32 +54,30 @@ class Inflector(ABC):
         raise NotImplementedError
 
     def get_verbs(self) -> List[Verb]:
-        return self._verb_parser.verbs
+        return list(self._verbs)
 
     def get_infinitives(self) -> List[str]:
-        return [v.infinitive for v in self._verb_parser.verbs]
+        return self._verbs.infinitives
 
     def get_templates(self) -> List[ConjugationTemplate]:
-        return self._conj_parser.templates
+        return list(self._conjugations)
 
     def get_template_names(self) -> List[str]:
-        return [t.name for t in self._conj_parser.templates]
+        return [t.name for t in self._conjugations]
 
     def find_verb_by_infinitive(self, infinitive: str) -> Verb:
-        return self._verb_parser.find_verb_by_infinitive(infinitive)
+        return self._verbs.find_verb_by_infinitive(infinitive)
 
     def find_template(self, name: str) -> ConjugationTemplate:
-        return self._conj_parser.find_template(name)
+        return self._conjugations.find_template(name)
 
     def get_verbs_that_start_with(self, query: str, max_results: int) -> List[str]:
-        query = query.lower()
-        matches = self._verb_parser.get_verbs_that_start_with(query, max_results)
-        return matches
+        return self._verbs.get_verbs_that_start_with(query.lower(), max_results)
 
     def get_verb_stem_from_template_name(
         self, infinitive: str, template_name: str
     ) -> str:
-        """Get the verb stem given an ininitive and a colon-delimited template name.
+        """Get the verb stem given an infinitive and a colon-delimited template name.
         E.g. infinitive='parler' template_name='aim:er' -> 'parl'
         Note: Catalan overrides this base class implementation to allow looser matching
         (only requires the last n-1 chars of template ending to match infinitive ending)