+++ verbecc/src/inflectors/lang/inflector_fr.py
@@ -60,7 +60,7 @@ class InflectorFr(Inflector):
     def get_verbs_that_start_with(self, query: str, max_results: int) -> List[str]:
         query = query.lower()
         is_reflexive, query = self.split_reflexive(query)
-        matches = self._verb_parser.get_verbs_that_start_with(query, max_results)
+        matches = self._verbs.get_verbs_that_start_with(query, max_results)
         if is_reflexive:
             matches = [
                 self.add_reflexive_pronoun(m)